+++ client/js/services/file-manager.service.js
@@ -21,7 +21,7 @@ import {
 } from "../pages/file-manager/file.template";
 import {
   generatePathList,
-  getFileListElement,
+  getFileListElement
 } from "../pages/file-manager/file-manager.helpers";
 
 const fileUploads = new Queue();
@@ -50,8 +50,13 @@ const handleUploadFromQueue = async (current, index) => {
     file.index = current.index;
     file.uploadDate = currentDate.getTime() / 1000;
     fileUploads.remove();
-    await OctoFarmClient.post("printers/newFiles", file);
-    await FileManagerSortingService.loadSort(current.index);
+    const fileUpdate = await OctoFarmClient.post("printers/newFiles", file);
+    if(!!fileUpdate?.files){
+      await FileManagerSortingService.loadSort(current.index, false, fileUpdate.files);
+    }else{
+      UI.createAlert("error", "File upload failed!", 3000, "clicked")
+    }
+
     const uploadsRemaining = document.getElementById("uploadsRemaining");
     if (uploadsRemaining) {
       uploadsRemaining.innerHTML = `${fileUploads.size()}`;
@@ -414,33 +419,6 @@ export default class FileManagerService {
     await FileManagerSortingService.loadSort(printer._id);
   }
 
-  static async openFolder(folder, target, printer) {
-    const fileBackButtonElement = document.getElementById("fileBackBtn");
-    if (typeof target !== "undefined" && target.type === "button") {
-      await FileManagerSortingService.loadSort(printer._id);
-      return;
-    }
-    if (typeof folder !== "undefined") {
-      folder = folder.replace("file-", "");
-
-      document.getElementById("currentFolder").innerHTML = `local/${folder}`;
-      fileBackButtonElement.disabled = false;
-    } else {
-      const currentFolder = document.getElementById("currentFolder").innerHTML;
-      if (currentFolder !== "local") {
-        const previousFolder = currentFolder.substring(
-          0,
-          currentFolder.lastIndexOf("/")
-        );
-        document.getElementById("currentFolder").innerHTML = previousFolder;
-        fileBackButtonElement.disabled = previousFolder === "local";
-      } else {
-        fileBackButtonElement.disabled = true;
-      }
-    }
-    await FileManagerSortingService.loadSort(printer._id);
-  }
-
   static async refreshFiles(printer, spinnerIcon) {
     if (fileUploads.size() <= 1) {
       for (const file of printer.fileList.fileList) {
@@ -507,93 +485,7 @@ export default class FileManagerService {
     }
   }
 
-  static updatePrinterMetrics(id, fileList, folderList) {
-    let currentFolder = document.getElementById("currentFolder").innerHTML;
-    if (currentFolder.includes("local/")) {
-      currentFolder = currentFolder.replace("local/", "");
-    }
 
-    const printerFileCount = document.getElementById("printerFileCount");
-    if (printerFileCount) {
-      printerFileCount.innerHTML = `<i class="fas fa-file"></i> ${fileList.length} <i class="fas fa-folder"></i> ${folderList.length}`;
-    }
-    const fileCardCount = document.getElementById("fileManagerFileCount-" + id);
-    if (fileCardCount) {
-      fileCardCount.innerHTML = `Files: ${fileList.length}`;
-    }
-
-    const fileCardFolderCount = document.getElementById(
-      "fileManagerFolderCount-" + id
-    );
-    if (fileCardFolderCount) {
-      fileCardFolderCount.innerHTML = `Folders: ${folderList.length}`;
-    }
-
-    return currentFolder;
-  }
-
-  static updatePrinterFilesList(printer, recursive) {
-    const { fileList } = printer;
-    // Update page elements
-    const currentFolder = FileManagerService.updatePrinterMetrics(
-      printer._id,
-      fileList.fileList,
-      fileList.folderList
-    );
-    //Draw folders,
-    if (!recursive) {
-      FileManagerService.drawFolders(
-        printer._id,
-        fileList.folderList,
-        currentFolder
-      );
-    }
-
-    // Draw Files,
-    FileManagerService.drawFiles(
-      printer._id,
-      fileList.fileList,
-      printer.printerURL,
-      currentFolder,
-      recursive
-    );
-
-    // Update Listeners
-    FileManagerService.updateListeners(printer);
-  }
-
-  static drawFolders(id, folderList, currentFolder) {
-    // Draw sub - folders present in current folder
-    const fileElem = getFileListElement(id);
-    if (folderList.length > 0) {
-      folderList.forEach((folder) => {
-        if (folder.path === currentFolder) {
-          fileElem.insertAdjacentHTML("beforeend", getFolderTemplate(folder));
-        }
-      });
-    }
-  }
-
-  static drawFiles(id, fileList, printerURL, currentFolder, recursive) {
-    const fileElem = getFileListElement(id);
-    if (fileElem) {
-      // Filter out files out of current folder scope
-      const currentFileList = fileList.filter((f) => {
-        return typeof recursive !== "undefined" || f.path === currentFolder;
-      });
-      // Show empty or filled list
-      if (currentFileList.length > 0) {
-        currentFileList.forEach((file) => {
-          fileElem.insertAdjacentHTML(
-            "beforeend",
-            getFileTemplate(file, printerURL, id)
-          );
-        });
-      } else {
-        fileElem.insertAdjacentHTML("beforeend", noFilesToShow());
-      }
-    }
-  }
 
   static async search(id) {
     await FileActions.search(id);
@@ -601,36 +493,6 @@ export default class FileManagerService {
 
   static async createFolder(printer) {
     await FileActions.createFolder(printer);
-  }
-
-  static updateListeners(printer) {
-    const fileElem = document.getElementById(`fileList-${printer._id}`);
-    dragAndDropEnableMultiplePrinters(fileElem, printer);
-    const folders = document.querySelectorAll(".folderAction");
-    folders.forEach((folder) => {
-      folder.addEventListener("click", async (e) => {
-        const updatedPrinter = await OctoFarmClient.getPrinter(printer._id);
-        await FileManagerService.openFolder(
-          folder.id,
-          e.target,
-          updatedPrinter
-        );
-      });
-    });
-    const fileActionBtns = document.querySelectorAll("[id*='*fileAction']");
-    fileActionBtns.forEach((btn) => {
-      // Gate Keeper listener for file action buttons
-      btn.addEventListener("click", async () => {
-        await FileManagerService.addCardListeners(printer, btn);
-      });
-    });
-    const folderActionBtns = document.querySelectorAll("[id*='*folderAction']");
-    folderActionBtns.forEach((btn) => {
-      // Gate Keeper listener for file action buttons
-      btn.addEventListener("click", async () => {
-        await FileManagerService.addCardListeners(printer, btn);
-      });
-    });
   }
 
   static async addCardListeners(printer, btn) {
