+++ server/services/server-side-events.service.js
@@ -52,14 +52,14 @@ const removeClient = (id) => {
 };
 
 const listActiveClients = (req, res) => {
-  const curretnActiveClientList = clientList.map((client) => {
+  const currentActiveClientList = clientList.map((client) => {
     return {
       connectionDate: client.id,
       userName: client?.user?.name ? client.user.name : UNKNOWN_USER,
       group: client.user.group
     };
   });
-  return res.json(curretnActiveClientList);
+  return res.json(currentActiveClientList);
 };
 /**
  *
